test-coloring: added an implementation.
This program reads an adjacency matrix and the colors assigned to each
node in a graph G=(V,E). It checks whether the coloring is good, that is,
if (u,v) \in E, then c(u) != c(v), where the function c:E->N represents the
color c(v) of node v.